feat:add pageable and sorted to Idioma

+++ src/main/java/com/sise/biblioteca/service/impl/IdiomaServiceImpl.java
@@ -1,16 +1,15 @@
 package com.sise.biblioteca.service.impl;
 
 import com.sise.biblioteca.entities.Idioma;
+import com.sise.biblioteca.errors.ClientException;
 import com.sise.biblioteca.repository.IIdiomaRepository;
 import com.sise.biblioteca.service.IIdiomaService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
-
 import org.springframework.data.domain.Pageable;
+import org.springframework.http.HttpStatus;
 import org.springframework.data.domain.Page;
-
-
 
 @Service
 public class IdiomaServiceImpl implements IIdiomaService {
@@ -20,12 +19,15 @@ public class IdiomaServiceImpl implements IIdiomaService {
 
   @Override
   public Page<Idioma> getAll(Pageable pageable) {
-    return idiomaRepository.findByEstado(true,pageable);
+    return idiomaRepository.findByEstado(true, pageable);
   }
 
   @Override
-  public Idioma getById(Integer idIdioma) {
-    return idiomaRepository.findOneByIdIdiomaAndEstado(idIdioma, true);
+  public Idioma getById(Integer idIdioma) throws ClientException {
+    Idioma idioma = idiomaRepository.findOneByIdIdiomaAndEstado(idIdioma, true);
+    if (idioma == null)
+      throw new ClientException("El idioma no existe", HttpStatus.NOT_FOUND);
+    return idioma;
   }
 
   @Override
@@ -35,12 +37,19 @@ public class IdiomaServiceImpl implements IIdiomaService {
   }
 
   @Override
-  public Idioma edit(Idioma idioma) {
-    return idiomaRepository.save(idioma);
+  public Idioma edit(Integer id, Idioma newIdioma) throws ClientException {
+    Idioma idioma = idiomaRepository.findOneByIdIdiomaAndEstado(id, true);
+    if (idioma == null)
+      throw new ClientException("El idioma no existe", HttpStatus.NOT_FOUND);
+    idioma = idiomaRepository.save(newIdioma);
+    return idioma;
   }
 
   @Override
-  public void remove(Integer idIdioma) {
+  public void remove(Integer idIdioma) throws ClientException {
+    Idioma idioma = idiomaRepository.findOneByIdIdiomaAndEstado(idIdioma, true);
+    if (idioma == null)
+      throw new ClientException("El idioma no existe", HttpStatus.NOT_FOUND);
     idiomaRepository.remove(idIdioma);
   }
